Initial commit.
Create a virtual LED and a group of LEDs to be controlled by it.
The virtual LED behaves like any other LED, so it can have a trigger
assigned to it etc.
The brightness of this LED is monitored, and the LEDs in the group are
set to the brightness of that parent LED when the brightness changes.
Useful in cases where multiple LEDs should change brightness together
(e.g. RGB LEDs where the colour is a composition of more than one LED.